rework Scan360, move to separate file
