+++ Assets/Scripts/Features/Scan360.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: e0ae523b75e2361479953ec139734116
+timeCreated: 1497188383
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
